chore(arch): Move parser to parser/Parser.ts
This isn't a big deal from an architectural perspective, but it's
helpful for new contributors.  It also makes it nice and clean to export
for consumers and import from unit tests!
