+++ test/parser/expression/Exponential.test.js
@@ -1,6 +1,5 @@
-const Parser = require("../../../lib/parser");
 const BrsError = require("../../../lib/Error");
-const { Lexeme, BrsTypes } = require("brs");
+const { Lexeme, BrsTypes, Parser } = require("brs");
 const { Int32 } = BrsTypes;
 
 const { EOF } = require("../ParserTests");
