+++ test/parser/statement/Set.test.js
@@ -1,7 +1,6 @@
-const Parser = require("../../../lib/parser");
 const BrsError = require("../../../lib/Error");
 
-const { Lexeme, BrsTypes } = require("brs");
+const { Lexeme, BrsTypes, Parser } = require("brs");
 const { Int32, BrsString } = BrsTypes;
 
 const { EOF } = require("../ParserTests");
